Added support for Bitcoin Cash
